+++ api/src/service/workflowitem_assign.ts
@@ -2,12 +2,14 @@ import { Ctx } from "../lib/ctx";
 import * as Result from "../result";
 import * as Cache from "./cache2";
 import { ConnToken } from "./conn";
+import { BusinessEvent } from "./domain/business_event";
 import { Identity } from "./domain/organization/identity";
 import { ServiceUser } from "./domain/organization/service_user";
 import * as Project from "./domain/workflow/project";
 import * as Subproject from "./domain/workflow/subproject";
 import * as Workflowitem from "./domain/workflow/workflowitem";
 import * as WorkflowitemAssign from "./domain/workflow/workflowitem_assign";
+import * as TypeEvents from "./domain/workflowitem_types/apply_workflowitem_type";
 import * as GroupQuery from "./group_query";
 import { store } from "./store";
 
@@ -34,6 +36,9 @@ export async function assignWorkflowitem(
         },
         getUsersForIdentity: async identity => {
           return GroupQuery.resolveUsers(conn, ctx, serviceUser, identity);
+        },
+        applyWorkflowitemType: (event: BusinessEvent, workflowitem: Workflowitem.Workflowitem) => {
+          return TypeEvents.applyWorkflowitemType(event, ctx, serviceUser, workflowitem);
         },
       },
     );