+++ api/src/service/workflowitem_close.ts
@@ -2,11 +2,13 @@ import { Ctx } from "../lib/ctx";
 import * as Result from "../result";
 import * as Cache from "./cache2";
 import { ConnToken } from "./conn";
+import { BusinessEvent } from "./domain/business_event";
 import { ServiceUser } from "./domain/organization/service_user";
 import * as Project from "./domain/workflow/project";
 import * as Subproject from "./domain/workflow/subproject";
 import * as Workflowitem from "./domain/workflow/workflowitem";
 import * as WorkflowitemClose from "./domain/workflow/workflowitem_close";
+import * as TypeEvents from "./domain/workflowitem_types/apply_workflowitem_type";
 import * as GroupQuery from "./group_query";
 import { store } from "./store";
 
@@ -28,6 +30,9 @@ export async function closeWorkflowitem(
       },
       getUsersForIdentity: async identity => {
         return GroupQuery.resolveUsers(conn, ctx, serviceUser, identity);
+      },
+      applyWorkflowitemType: (event: BusinessEvent, workflowitem: Workflowitem.Workflowitem) => {
+        return TypeEvents.applyWorkflowitemType(event, ctx, serviceUser, workflowitem);
       },
     }),
   );