+++ api/src/service/workflowitem_create.ts
@@ -2,10 +2,13 @@ import { Ctx } from "../lib/ctx";
 import * as Result from "../result";
 import * as Cache from "./cache2";
 import { ConnToken } from "./conn";
+import { BusinessEvent } from "./domain/business_event";
 import { ServiceUser } from "./domain/organization/service_user";
 import { ResourceMap } from "./domain/ResourceMap";
-import * as Workflowitem from "./domain/workflow/workflowitem_create";
+import * as Workflowitem from "./domain/workflow/workflowitem";
+import * as WorkflowitemCreate from "./domain/workflow/workflowitem_create";
 import * as WorkflowitemCreated from "./domain/workflow/workflowitem_created";
+import * as TypeEvents from "./domain/workflowitem_types/apply_workflowitem_type";
 import { store } from "./store";
 
 export { RequestData } from "./domain/workflow/workflowitem_create";
@@ -14,10 +17,10 @@ export async function createWorkflowitem(
   conn: ConnToken,
   ctx: Ctx,
   serviceUser: ServiceUser,
-  requestData: Workflowitem.RequestData,
+  requestData: WorkflowitemCreate.RequestData,
 ): Promise<ResourceMap> {
-  const result = await Cache.withCache(conn, ctx, (cache) => {
-    return Workflowitem.createWorkflowitem(ctx, serviceUser, requestData, {
+  const result = await Cache.withCache(conn, ctx, cache => {
+    return WorkflowitemCreate.createWorkflowitem(ctx, serviceUser, requestData, {
       workflowitemExists: async (
         projectId: string,
         subprojectId: string,
@@ -28,6 +31,9 @@ export async function createWorkflowitem(
       },
       getSubproject: async (projectId: string, subprojectId: string) =>
         await cache.getSubproject(projectId, subprojectId),
+      applyWorkflowitemType: (event: BusinessEvent, workflowitem: Workflowitem.Workflowitem) => {
+        return TypeEvents.applyWorkflowitemType(event, ctx, serviceUser, workflowitem);
+      },
     });
   });
 
@@ -38,7 +44,7 @@ export async function createWorkflowitem(
     await store(conn, ctx, event);
   }
 
-  const workflowitemEvent = newEvents.find((x) => (x as any).workflowitem.id !== undefined);
+  const workflowitemEvent = newEvents.find(x => (x as any).workflowitem.id !== undefined);
   if (workflowitemEvent === undefined) throw Error(`Assertion: This is a bug.`);
 
   const resourceIds: ResourceMap = {