+++ frontend/src/pages/Workflows/workflowitemTypes.js
@@ -1,0 +1,8 @@
+import strings from "../../localizeStrings";
+
+export const types = ["general", "restricted"];
+
+export const typesDescription = {
+  general: strings.workflow.workflowitem_type_general,
+  restricted: strings.workflow.workflowitem_type_restricted
+};